requirements.txt
streamlit
langchain-huggingface
HuggingFaceEndpoint
langchain
python-dotenv
huggingface_hub